+++ Changelog.md
@@ -72,6 +72,8 @@ Note: Numbers starting with a "#" like #330 refer to the bugreport with that num
 * Improved compatibility with Wayland (#426)
 * Work around assertion in AlphaLabs4 due to "ride_of_death" yeeting
   the dead "monster_zsec_shotgun_12" into the void (#409)
+* Support loading some mods known to need `fs_game_base d3xp` via Mods menu
+  (currently, *The Lost Mission* and *LibreCoop d3xp* are supported)
 
 1.5.1 (2021-03-14)
 ------------------------------------------------------------------------